debian: Make it easy to build a debian source package
The actual debian files are from pharo-ubuntu and Norbert's
Pharo5 packaging.

The approach is to build the package in the same repository as
the VM. This way one doesn't need to wait for the sources to be
published and can add travis.ci tests as well. For the Ubuntu
n-version packaging a solution must be found. For the 64bit build
we probably generate a second template.

Patch the resulting cmake file to not download libssh2/libgit as
they are not utilized anyway and libSDL2 as it needs to come from
the system for Debian/Ubuntu packages.

Complete the pharo5-vm renaming and update pharo5-vm-x and the
desktop files.